semester-03/software-dev-project/week-08/module-30/drf_part_2/smart_care/patient/urls.py
from rest_framework.routers import DefaultRouter
from django.urls import include,path
from .views import PatientViewSet

router=DefaultRouter()
router.register('list',PatientViewSet)

urlpatterns = [
    path('',include(router.urls)),
]
